refactor: component list keys, code cleanup
