+++ src/components/SearchResults/SearchResults.js
@@ -46,8 +46,8 @@ const SearchResults = ({ isSearching, searchResults, keyword }) => {
         <ArticleSection>
           {displaySets !== 0 && <h2 className='results__title'>Search results</h2>}
           <ul className='results__article-container'>
-            {displayCards.map((card, i) => (
-              <NewsCard key={i} keyword={keyword} {...card}></NewsCard>
+            {displayCards.map((card) => (
+              <NewsCard key={card.author + card.source.name + card.publishedAt} keyword={keyword} {...card}></NewsCard>
             ))}
           </ul>
           {!isSearching && displayCards.length < searchResults.length && <ShowMoreButton getNextCards={handleGetNextCards} />}
