+++ src/utils/constants.js
@@ -1,0 +1,1 @@
+export const MAX_MOBILE_SIZE = 650;
